pyproject.toml
[tool.poetry]
name = "hack-api"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"
packages = [{include = "hack_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.87.0"
sqlalchemy = "^1.4.44"
pytest = "^7.2.0"
factory-boy = "^3.2.1"
pytest-freezegun = "^0.4.2"
asyncpg = "^0.27.0"
greenlet = "^2.0.1"
lxml = "^4.9.1"
sqlalchemy-utils = "^0.38.3"
structlog = "^22.3.0"
uvloop = "^0.17.0"
sqladmin = "^0.8.0"
gunicorn = "^20.1.0"
levenshtein = "^0.20.9"
requests = "^2.28.1"
bs4 = "^0.0.1"
python-dotenv = "^0.21.0"
typer = "^0.7.0"
questionary = "^1.10.0"
rodi = "^1.1.3"
uvicorn = "^0.20.0"
dependency-injector = "^4.41.0"
asgi-lifespan = "^2.0.0"
pytest-xdist = "^3.1.0"
psycopg2-binary = "^2.9.5"
rich = "^13.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
